Liberty Ridge Update, Sponsors, Calendar
Updated a number of pages with this commit. The Calendar page now refers to the
public URL for the Basecamp Schedule - our primary calendar now. The sign-up
link was removed, soon to be replaced. The About page needs more work. A new
Sponsors page was added. Media to come soon.